refactor(auth): move ORM models into auth_v2
