+++ backend/core/auth/models/auth_session.py
@@ -1,42 +1,3 @@
-import uuid
-from datetime import datetime, timezone
+from core.auth_v2.models.auth_session import AuthSession
 
-from sqlalchemy import DateTime, ForeignKey, String
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import Mapped, mapped_column
-
-from core.common import ORMBase
-
-
-class AuthSession(ORMBase):
-    __tablename__ = "session"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    user_id: Mapped[uuid.UUID] = mapped_column(
-        ForeignKey("user.id", ondelete="CASCADE")
-    )
-    refresh_token_hash: Mapped[str] = mapped_column(String)
-    user_agent: Mapped[str | None] = mapped_column(String(512), nullable=True)
-    ip: Mapped[str | None] = mapped_column(String(45), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-    expires_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False
-    )
-    revoked_at: Mapped[datetime | None] = mapped_column(
-        DateTime(timezone=True), nullable=True
-    )
-    replaced_by_session_id: Mapped[uuid.UUID | None] = mapped_column(
-        UUID(as_uuid=True), ForeignKey("session.id"), nullable=True
-    )
-    last_used_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True),
-        default=lambda: datetime.now(timezone.utc),
-        onupdate=lambda: datetime.now(timezone.utc),
-    )
-
-    def __repr__(self) -> str:
-        return f"Session(id={self.id}, user_id={self.user_id}, created_at={self.created_at}, expires_at={self.expires_at}"
+__all__ = ["AuthSession"]
