+++ backend/core/auth/models/user.py
@@ -1,30 +1,3 @@
-import uuid
-from datetime import datetime, timezone
+from core.auth_v2.models.user import User
 
-from sqlalchemy import DateTime
-from sqlalchemy.dialects.postgresql import JSONB, UUID
-from sqlalchemy.orm import Mapped, mapped_column
-
-from core.common import ORMBase
-
-
-class User(ORMBase):
-    __tablename__ = "user"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    email: Mapped[str] = mapped_column(unique=True)
-    password_hash: Mapped[str]
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True),
-        default=lambda: datetime.now(timezone.utc),
-        onupdate=lambda: datetime.now(timezone.utc),
-    )
-    meta: Mapped[dict] = mapped_column(JSONB, default=dict)
-
-    def __repr__(self) -> str:
-        return f"User(email={self.email}, id={self.id})"
+__all__ = ["User"]
