+++ backend/core/auth_v2/__init__.py
@@ -1,0 +1,3 @@
+from . import models
+
+__all__ = ["models"]
